move to ShowWindow testcase
