+++ modules/rostests/apitests/user32/ShowWindow.c
@@ -2,10 +2,11 @@
  * PROJECT:     ReactOS API tests
  * LICENSE:     LGPL-2.1+ (https://spdx.org/licenses/LGPL-2.1+)
  * PURPOSE:     Tests for ShowWindow
- * COPYRIGHT:   Copyright 2021 Katayama Hirofumi MZ <katayama.hirofumi.mz@gmail.com>
+ * COPYRIGHT:   Copyright 2021-2025 Katayama Hirofumi MZ <katayama.hirofumi.mz@gmail.com>
  */
 
 #include "precomp.h"
+#include <versionhelpers.h>
 
 typedef struct TEST_ENTRY
 {
@@ -164,7 +165,7 @@ WindowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
     return 0;
 }
 
-START_TEST(ShowWindow)
+static VOID Test_ShowWindow_Main(VOID)
 {
     WNDCLASSA wc;
     UINT iTest;
@@ -180,4 +181,165 @@ START_TEST(ShowWindow)
     {
         DoTestEntry(&s_entries[iTest]);
     }
+}
+
+#define RED RGB(255, 0, 0)
+
+static COLORREF CheckColor(VOID)
+{
+    HDC hDC = GetDC(NULL);
+    COLORREF color = GetPixel(hDC, 100, 100);
+    ReleaseDC(NULL, hDC);
+    return color;
+}
+
+static DWORD WINAPI
+ForceMinimizeThreadFunc(LPVOID arg)
+{
+    BOOL ret;
+    HWND hwnd = (HWND)arg;
+    DWORD style, exstyle;
+
+    Sleep(100);
+    ok_long(CheckColor(), RED);
+
+    SetWindowLongPtrW(hwnd, GWL_EXSTYLE, WS_EX_MAKEVISIBLEWHENUNGHOSTED);
+    ret = ShowWindow(hwnd, SW_FORCEMINIMIZE);
+    Sleep(100);
+    ok(ret != FALSE, "ret was FALSE\n");
+    ok(CheckColor() != RED, "Color was red\n");
+    style = GetWindowLongPtrW(hwnd, GWL_STYLE);
+    if (IsWindowsVistaOrGreater())
+        ok((style & (WS_POPUP | WS_MINIMIZE | WS_VISIBLE)) == (WS_POPUP | WS_MINIMIZE | WS_VISIBLE), "style was 0x%08lX\n", style);
+    else
+        ok((style & (WS_POPUP | WS_MINIMIZE | WS_VISIBLE)) == (WS_POPUP), "style was 0x%08lX\n", style);
+    exstyle = GetWindowLongPtrW(hwnd, GWL_EXSTYLE);
+    ok_long(exstyle, 0);
+
+    ShowWindow(hwnd, SW_MINIMIZE);
+    Sleep(100);
+    ok(CheckColor() != RED, "Color was red\n");
+
+    SetWindowLongPtrW(hwnd, GWL_EXSTYLE, WS_EX_MAKEVISIBLEWHENUNGHOSTED);
+    ret = ShowWindow(hwnd, SW_FORCEMINIMIZE);
+    Sleep(100);
+    ok(ret != FALSE, "ret was FALSE\n");
+    ok(CheckColor() != RED, "Color was red\n");
+    style = GetWindowLongPtrW(hwnd, GWL_STYLE);
+    ok((style & (WS_POPUP | WS_MINIMIZE | WS_VISIBLE)) == (WS_POPUP | WS_MINIMIZE | WS_VISIBLE), "style was 0x%08lX\n", style);
+    exstyle = GetWindowLongPtrW(hwnd, GWL_EXSTYLE);
+    ok_long(exstyle, 0);
+
+    ShowWindow(hwnd, SW_SHOWNORMAL);
+    Sleep(100);
+    ok_long(CheckColor(), RED);
+
+    SetWindowLongPtrW(hwnd, GWL_EXSTYLE, WS_EX_MAKEVISIBLEWHENUNGHOSTED);
+    ret = ShowWindow(hwnd, SW_FORCEMINIMIZE);
+    Sleep(100);
+    ok(ret != FALSE, "ret was FALSE\n");
+    ok(CheckColor() != RED, "Color was red\n");
+    style = GetWindowLongPtrW(hwnd, GWL_STYLE);
+    if (IsWindowsVistaOrGreater())
+        ok((style & (WS_POPUP | WS_MINIMIZE | WS_VISIBLE)) == (WS_POPUP | WS_MINIMIZE | WS_VISIBLE), "style was 0x%08lX\n", style);
+    else
+        ok((style & (WS_POPUP | WS_MINIMIZE | WS_VISIBLE)) == (WS_POPUP), "style was 0x%08lX\n", style);
+    exstyle = GetWindowLongPtrW(hwnd, GWL_EXSTYLE);
+    ok_long(exstyle, 0);
+
+    SetWindowLongPtrW(hwnd, GWL_EXSTYLE, WS_EX_MAKEVISIBLEWHENUNGHOSTED);
+    ret = ShowWindow(hwnd, SW_FORCEMINIMIZE);
+    Sleep(100);
+    if (IsWindowsVistaOrGreater())
+        ok(ret != FALSE, "ret was FALSE\n");
+    else
+        ok_bool_false(ret, "Return was");
+    ok(CheckColor() != RED, "Color was red\n");
+    style = GetWindowLongPtrW(hwnd, GWL_STYLE);
+    if (IsWindowsVistaOrGreater())
+        ok((style & (WS_POPUP | WS_MINIMIZE | WS_VISIBLE)) == (WS_POPUP | WS_MINIMIZE | WS_VISIBLE), "style was 0x%08lX\n", style);
+    else
+        ok((style & (WS_POPUP | WS_MINIMIZE | WS_VISIBLE)) == (WS_POPUP), "style was 0x%08lX\n", style);
+    exstyle = GetWindowLongPtrW(hwnd, GWL_EXSTYLE);
+    ok_long(exstyle, 0);
+
+    PostMessageW(hwnd, WM_CLOSE, 0, 0);
+    return 0;
+}
+
+static LRESULT CALLBACK
+ForceMinimizeWndProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
+{
+    switch (uMsg)
+    {
+        case WM_DESTROY:
+            PostQuitMessage(0);
+            break;
+        default:
+            return DefWindowProcW(hwnd, uMsg, wParam, lParam);
+    }
+    return 0;
+}
+
+static VOID
+Test_WS_FORCEMINIMIZE_Sub(HBRUSH hbr)
+{
+    WNDCLASSW wc;
+    HWND hwnd;
+    DWORD style;
+    HINSTANCE hInstance = GetModuleHandleW(NULL);
+    DWORD dwThreadId;
+    HANDLE hThread;
+    MSG msg;
+
+    ZeroMemory(&wc, sizeof(wc));
+    wc.lpfnWndProc = ForceMinimizeWndProc;
+    wc.hInstance = hInstance;
+    wc.hIcon = LoadIconW(NULL, (PCWSTR)IDI_APPLICATION);
+    wc.hCursor = LoadCursorW(NULL, (PCWSTR)IDC_ARROW);
+    wc.hbrBackground = hbr;
+    wc.lpszClassName = L"SW_FORCEMINIMIZE";
+    if (!RegisterClassW(&wc))
+    {
+        skip("RegisterClassW failed\n");
+        return;
+    }
+
+    style = WS_POPUP | WS_VISIBLE;
+    hwnd = CreateWindowExW(0, L"SW_FORCEMINIMIZE", L"SW_FORCEMINIMIZE", style,
+                           50, 50, 100, 100, NULL, NULL, hInstance, NULL);
+    if (!hwnd)
+    {
+        skip("CreateWindowExW failed\n");
+        return;
+    }
+
+    hThread = CreateThread(NULL, 0, ForceMinimizeThreadFunc, hwnd, 0, &dwThreadId);
+    if (!hThread)
+    {
+        skip("CreateThread failed\n");
+        DestroyWindow(hwnd);
+        return;
+    }
+    CloseHandle(hThread);
+
+    while (GetMessageW(&msg, NULL, 0, 0))
+    {
+        TranslateMessage(&msg);
+        DispatchMessageW(&msg);
+    }
+}
+
+static VOID
+Test_WS_FORCEMINIMIZE()
+{
+    HBRUSH hbr = CreateSolidBrush(RED);
+    Test_WS_FORCEMINIMIZE_Sub(hbr);
+    DeleteObject(hbr);
+}
+
+START_TEST(ShowWindow)
+{
+    Test_ShowWindow_Main();
+    Test_WS_FORCEMINIMIZE();
 }
